pyproject.toml
[project]
name = "pi5ip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord>=2.3.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
